Refactor to use API in visual mode
