+++ librz/core/core_private.h
@@ -6,6 +6,10 @@
 
 RZ_IPI int rz_core_analysis_set_reg(RzCore *core, const char *regname, ut64 val);
 RZ_IPI void rz_core_analysis_esil_init(RzCore *core);
+RZ_IPI bool rz_core_analysis_var_rename(RzCore *core, const char *name, const char *newname);
+
+/* cdebug.c */
+RZ_IPI bool rz_core_debug_reg_set(RzCore *core, const char *regname, ut64 val, const char *strval);
 RZ_IPI bool rz_core_debug_reg_list(RzCore *core, int type, int size, PJ *pj, int rad, const char *use_color);
 RZ_IPI void rz_core_debug_regs2flags(RzCore *core, int bits);
 RZ_IPI void rz_core_regs2flags(RzCore *core);
