+++ librz/core/vmenus.c
@@ -1,8 +1,11 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 
+#include <string.h>
+
 #include <rz_core.h>
 #include <rz_util.h>
-#include <string.h>
+
+#include "core_private.h"
 
 #define MAX_FORMAT 3
 
@@ -1229,7 +1232,8 @@ RZ_API int rz_core_visual_classes(RzCore *core) {
 		case 'p':
 			if (mode == 'm' && mur) {
 				rz_core_seek(core, mur->vaddr, true);
-				rz_core_cmd0(core, "af;pdf~..");
+				rz_core_analysis_function_add(core, NULL, core->offset, false);
+				rz_core_cmd0(core, "pdf~..");
 			}
 			break;
 		case 'm': // methods
@@ -1745,8 +1749,7 @@ RZ_API int rz_core_visual_view_rop(RzCore *core) {
 			rz_line_set_prompt("comment: ");
 			const char *line = rz_line_readline();
 			if (line && *line) {
-				// XXX code injection bug here
-				rz_core_cmdf(core, "CC %s @ 0x%08" PFMT64x, line, addr + delta);
+				rz_meta_set_string(core->analysis, RZ_META_TYPE_COMMENT, addr + delta, line);
 			}
 		} break;
 		case '.':
@@ -2229,7 +2232,7 @@ RZ_API int rz_core_visual_comments(RzCore *core) {
 		case ' ':
 		case '\r':
 		case '\n':
-			rz_core_cmdf(core, "s 0x%" PFMT64x, from);
+			rz_core_seek_and_save(core, from, true);
 			RZ_FREE(p);
 			return true;
 		case 'Q':
@@ -2517,7 +2520,7 @@ static void variable_rename(RzCore *core, ut64 addr, int vindex, const char *nam
 	rz_list_foreach (list, iter, var) {
 		if (i == vindex) {
 			rz_core_seek(core, addr, false);
-			rz_core_cmd_strf(core, "afvn %s %s", name, var->name);
+			rz_core_analysis_var_rename(core, name, var->name);
 			rz_core_seek(core, a_tmp, false);
 			break;
 		}
@@ -2881,7 +2884,7 @@ RZ_API void rz_core_visual_debugtraces(RzCore *core, const char *input) {
 			rz_core_cmdf(core, ".dte %d", i);
 		}
 		rz_core_cmd0(core, "x 64@r:SP");
-		rz_core_cmd0(core, "dri");
+		rz_core_debug_ri(core, core->dbg->reg, 0);
 		// limit by rows here
 		//int rows = rz_cons_get_size (NULL);
 		rz_core_cmdf(core, "dtd %d", delta);
@@ -3045,7 +3048,12 @@ RZ_API void rz_core_visual_analysis(RzCore *core, const char *input) {
 		case 'a':
 			switch (level) {
 			case 0:
-				rz_core_cmd0(core, "af-$$;af"); // reanalize
+				// Remove the old function information
+				rz_core_analysis_undefine(core, core->offset);
+				rz_analysis_fcn_del_locs(core->analysis, core->offset);
+				rz_analysis_fcn_del(core->analysis, core->offset);
+				// Reanalyze and create function from scratch
+				rz_core_analysis_function_add(core, NULL, core->offset, false);
 				break;
 			case 1: {
 				eprintf("Select variable source ('r'egister, 's'tackptr or 'b'aseptr): ");
@@ -3127,7 +3135,10 @@ RZ_API void rz_core_visual_analysis(RzCore *core, const char *input) {
 		case '-':
 			switch (level) {
 			case 0:
-				rz_core_cmdf(core, "af- 0x%" PFMT64x, addr);
+				// Remove the old function information
+				rz_core_analysis_undefine(core, addr);
+				rz_analysis_fcn_del_locs(core->analysis, addr);
+				rz_analysis_fcn_del(core->analysis, addr);
 				break;
 			}
 			break;
@@ -3701,7 +3712,8 @@ onemoretime:
 			rz_analysis_function_resize(fcn, core->offset - fcn->addr);
 		}
 		rz_cons_break_push(NULL, NULL);
-		rz_core_cmdf(core, "af @ 0x%08" PFMT64x, off); // required for thumb autodetection
+		// required for thumb autodetection
+		rz_core_analysis_function_add(core, NULL, off, false);
 		rz_cons_break_pop();
 	} break;
 	case 'v': {
